pyproject.toml
[tool.poetry]
name = "notifypub"
version = "0.1.0"
description = "Notification Publisher"
authors = ["Isaac Beverly"]
license = "AGPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
waitress = "^2.1.2"
python-dotenv = "^0.21.0"
Flask-WTF = "^1.0.1"
Flask-Session = "^0.4.0"
Flask-Login = "^0.6.2"
SQLAlchemy = "^1.4.41"
pytest-flask = "^1.2.0"


[tool.poetry.group.dev.dependencies]
poetry = "^1.2.0"
pytest-flask = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
